test: split test-crypto-keygen.js
To avoid timing out on ARM machines in the CI.
